backend/requirements.txt
flask
flask-cors
yfinance
pandas
numpy
python-dotenv
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
yfinance==0.2.41
pandas==2.2.2
numpy==2.0.1